Modified the "days of the week" code to make the first day configurable. In OpenAir, the business is able to change the first day of the week. Previously this extension had the first day hardcoded to "Monday". In this change I add a new dropdown to the Options menu allowing the user to select the first day. It still defaults to Monday.
Since loading the options is asynchronous, we can no longer rely on "getDayNum()" to return the correct day until we have loaded this setting. That means we need to delay the timesheet code that depends on this function. I've moded that code to a callback after the "firstDayOfWeek" setting is loaded.